# Avoid allocations when using delay_serialization